feat(multi-segment-edges): extended polyline and bend point logic for inheritance edges
